Add basic p2p
There is a new socket ROUTER-DEALER connection among daemons. Besides
the existing REQ-REP socket, used to communicate with a client, the
ROUTER-DEALER is asynchronous on both send and recv sides and listen on
the RPC port, see command line arg --rpc-port.

An optional peer can also now be specified via the --peer command line
arg, to which the daemon will immediately connect to and which when it
receives the connection the peer daemon will connect back.

The CLI can now also query the number of peers connected to the daemon
it connects to.

Added some tests, exercising the handshake between two peer daemons,
initiated by the one that has the --peer argument specified, and CLI
querying peers on two running daemons.