Daemon now attempts to bind to a range of ports
If the default port is alrady taken, deamon now attempts to bind to the
next, and it gives up until 10 tries. This is done for both server and
rpc ports.

If the user explicitly specifies the ports on the command line, using
the -p or -r arguments, we only attempt to bind to the specific ports
given.

Since attempting to bind to different rpc ports may update the port that
is used and on another thread we need to send to peers what rpc port we
are bound, a std::condition_variable (with a std::mutex) is used to
ensure binding (and updating the port) happens before this information
is sent to peers.