Add broadcast of peers to daemon
This allows connecting peers that did not know about each other when
started. For example, if A and B start with B specifying A as its
initial peer, both A and B connect, as before. Previously if C also
started, it only connected to its initial peers, explicitly specified.
Now if C starts specifying only A as its initial peer, after connecting
to A, it will eventually also know about B and connect to it.